final-pjt-back/requirements.txt
asgiref==3.4.1
Django==3.2.6
django-appconf==1.0.5
django-cors-headers==3.14.0
django-imagekit==4.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
pilkit==2.0
Pillow==9.5.0
PyJWT==2.7.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
